Conclusão 20-09-2021

+++ ws/forum/src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -73,17 +73,30 @@ public class TopicosController {
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
-		Topico topico = form.atualizar(id, topicoRepository);
+		Optional<Topico> optional = topicoRepository.findById(id);
 		
-		return ResponseEntity.ok(new TopicoDto(topico));
+		if(optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+
+			return ResponseEntity.ok(new TopicoDto(topico));
+
+		}
+		
+		return ResponseEntity.notFound().build();		
 	}
 	
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id) {
-		topicoRepository.deleteById(id);
+		Optional<Topico> optional = topicoRepository.findById(id);
 		
-		return ResponseEntity.ok().build();
+		if(optional.isPresent()) {
+			topicoRepository.deleteById(id);		
+			return ResponseEntity.ok().build();
+		}
+		
+		return ResponseEntity.notFound().build();		
+
 	}
 
 }
